Only print ktrans output on error, or when not quiet.
Looks like console output makes ktrans.exe slow, which is especially
noticable when compiling large projects (lots of binary targets). This
change makes it possible to have ktransw suppress ktrans output if
configured to do so with the '--quiet' option.

Ktrans output is always shown if there was an error while translating.