Add support for cpp's '-MP' option.
This makes ktransw generate additional targets (with no prerequisites or
commands) that force make to rebuild all targets that depend on them if
they don't exist.

By doing this a header may be renamed (after generation of the dependency
files in a previous run) without having to manually update the (generated)
Makefile.

Technique described in "Autodependencies with GNU make" [1] by Scott
McPeak.

[1] http://scottmcpeak.com/autodepend/autodepend.html