Add pre-processor support.
This is a major improvement of ktransw, as by letting the pre-processor do
all the work of resolving, tracking and merging includes, there is no need
to create a temporary build directory that includes all the used include
files in the expected relative location.

The Karel source passed on to ktrans.exe is a completely stand-alone, flat
source file, which ktrans.exe can translate without needing any additional
external artefacts (other than the default 'support' and 'environment'
directories which it already knows how to find).

The pre-processor used -- gpp -- adds full support for C-like
pre-processor directives (if, if(n)def, define, etc, etc) and additionally
for macros. To fit better with other Karel source -- style-wise -- the
traditional '#' for pre-processor directives has been changed to '%', and
all directives are expected to be upper case.

Known limitations / changes / shortcomings:

 - include directives MUST use the full filename + extension (ie:
   'my_include.h.kl', instead of the 'my_include.h' that plain ktrans
   accepts). This will make sources using this style of includes
   incompatible with plain ktrans and Roboguide
 - plain ktrans chokes on unknown 'translator directives' (ie: '%DEFINE').
   This means that Karel source that uses this will no longer be
   translatable by plain ktrans and Roboguide
 - the 'support dir' for the selected Karel runtime version (ex: V7.70-1)
   must be explicitly added to the include path passed to ktransw, as
   includes are now handled by the pre-processor, not ktrans itself
   anymore

It may be possible to provide some sort of backward-compatible notation
for the newly added directives to allow plain ktrans to ignore them. That
may be added in a later revision.

It may also be possible for gpp to ignore a set of 'system headers', which
would remove the need for the 'support dir' to be added to the include
path of ktransw. This may also be added in a later revision.